backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
chromadb==0.4.15
numpy==1.24.3
pandas==2.0.3